+++ src/abstract/DropdownMenuItem.ts
@@ -1,4 +1,4 @@
-import { Component } from 'vue'
+import { Component, Ref } from 'vue'
 import { Maybe } from './BasicTypes'
 
 // The key must be assigned to let compare items automatically
@@ -6,5 +6,5 @@ export type DropdownMenuItemConstructor<I> = {
   widget: Component
   value?: Maybe<I>
   key: string
-  title: string
+  title: Ref<string>
 }